verifier/tdx: split quote verification into pluggable backends
Extract the ECDSA quote-verification step out of tdx::quote into a new
tdx::verify module so that the backend can be selected at compile time.
Today there is exactly one backend, tdx-dcap-ffi, which contains the
existing Intel DCAP QVL (libsgx_dcap_quoteverify) FFI implementation
moved verbatim; behaviour is unchanged.

The tdx-verifier feature no longer implies a backend: a backend feature
(currently only tdx-dcap-ffi) must be enabled, and enabling tdx-verifier
without one is a clear compile_error. all-verifier, az-tdx-vtpm-verifier
and the attestation-service passthroughs are wired to tdx-dcap-ffi, so
every default build keeps using the DCAP shared library exactly as before.

This is groundwork for a second, pure-Rust backend (tdx-dcap-rust).
